move errorHandler file to middleware folder
